Rename *-whitelist rules to *-allowlist
This applies a full rename across the repository to reflect that the
term "whitelist" is controversial in the tech community and that using
the term "allowlist" is more explicit in it's purpose.

In a later commit I'm going to reintroduce the whitelist rules as deprecated
rules. I felt this approach was preferred to creating new rules for the
allowlist rules as this allowed them to maintain their history.