Set up a rule factory in allowlist/denylist rules
This is to provide a function that can generate the rule function. The
purpose of this is to allow all of these rules to be used again in their
whitelist/blacklist form so that code does not need to be duplicated.
The only difference for whitelist/blacklist will be that they provide a
deprecated warning.